+++ lib/karafka/web/config.rb
@@ -48,7 +48,8 @@ module Karafka
             Tracking::Consumers::Listeners::Statistics.new,
             Tracking::Consumers::Listeners::Pausing.new,
             Tracking::Consumers::Listeners::Processing.new,
-            Tracking::Consumers::Listeners::Tags.new
+            Tracking::Consumers::Listeners::Tags.new,
+            Tracking::Consumers::Listeners::Connections.new
           ]
         end